* m4/compiler.m4: Add -Wextra, but turn off some of the silly ones. * nih/io.c (nih_io_buffer_resize): Remove unnecessary assert. * nih/file.c (nih_file_read, nih_file_map): Be explicit when comparing signed and unsigned types. * nih/option.c (nih_option_long): Cast pointer difference to size_t (nih_option_group_help): Use unsigned constant. * nih/command.c (nih_command_group_help): Use unsigned constant. * nih/tests/test_config.c (test_parse_stanza): Cast -1 to size_t to get the real number. * nih-dbus/dbus_object.c (nih_dbus_object_vtable): Move static to before const. * nih-dbus/tests/test_dbus_message.c (my_error_connect): Put static before const.

+++ nih/option.c
@@ -384,7 +384,7 @@ nih_option_long (NihOptionCtx *ctx)
 	 */
 	arg = ctx->argv[ctx->arg] + 2;
 	ptr = strchr (arg, '=');
-	len = (ptr ? ptr - arg : strlen (arg));
+	len = (ptr ? (size_t)(ptr - arg) : strlen (arg));
 
 	/* Find the option */
 	opt = nih_option_get_long (ctx, arg, len);
@@ -940,7 +940,7 @@ nih_option_group_help (NihOptionGroup  *group,
 		printf (_("Options:\n"));
 	}
 
-	width = nih_max (nih_str_screen_width (), 50) - 31;
+	width = nih_max (nih_str_screen_width (), 50U) - 31;
 
 	for (opt = options; (opt->option || opt->long_option); opt++) {
 		nih_local char *str = NULL;
